AG - Velocidade e Acelerações ajustadas para double, normalização da entrada e saída da RN

+++ semestre_04/alg_evolutivos/projeto_ag/components/cannon.cpp
@@ -25,7 +25,7 @@ int Cannon::get_enemy_distance(const Enemy &e){
     return e.pos_x - this->pos_x;
 }
 
-Projectile Cannon::shot_projectile(int vel_x, int vel_y, int degrees){
+Projectile Cannon::shot_projectile(double vel_x, double vel_y, int degrees){
     int pos_x = this->pos_x + (this->width/2);
     int pos_y = this->pos_y + (this->height/2);
     int radius = 10;
